+++ packages/types/src/index.ts
@@ -9,5 +9,3 @@
  */
 export * from './core/base'
 export * from './core/option'
-export * from './performance/FMP/base'
-export * from './performance/FMP/option'
