--- packages/types/src/performance/FMP/base.ts
@@ -1,8 +1,0 @@
-export interface FMPRecodeData {
-  time: number
-  domScore: number
-}
-
-export type HTMLElementWithCss = HTMLElement & {
-  readonly style?: CSSStyleDeclaration
-}
