--- packages/types/src/performance/FMP/option.ts
@@ -1,3 +1,0 @@
-export interface FMPOptions {
-    exact?: boolean
-}
